+++ packages/tailwindcss/src/utils/decode-arbitrary-value.test.ts
@@ -93,6 +93,8 @@ describe('adds spaces around math operators', () => {
       'radial-gradient(calc(1+2)),radial-gradient(calc(1+2))',
       'radial-gradient(calc(1 + 2)),radial-gradient(calc(1 + 2))',
     ],
+    ['w-[calc(anchor-size(width)+8px)]', 'w-[calc(anchor-size(width) + 8px)]'],
+    ['w-[calc(anchor-size(foo(bar))+8px)]', 'w-[calc(anchor-size(foo(bar)) + 8px)]'],
 
     [
       '[content-start]_calc(100%-1px)_[content-end]_minmax(1rem,1fr)',